+++ src/compiler.cpp
@@ -256,6 +256,10 @@ int Compiler::operator()(ObjPtr<FuncObject> &func) {
   this->frontEnd.teardownASTDump();
   assert(this->frontEnd.getContext().size() == 1);
   {
+    auto *prevType = this->frontEnd.getPrevType();
+    if (prevType && prevType->isNothingType()) {
+      this->frontEnd.getContext().back()->scope->updateModAttr(ModAttr::UNREACHABLE);
+    }
     auto &modType = hasFlag(this->compileOption, CompileOption::SINGLE_EXPR)
                         ? this->provider.getPool().getBuiltinModType()
                         : this->provider.newModTypeFromCurContext(this->frontEnd.getContext());