+++ test/analyzer/archive_test.cpp
@@ -46,7 +46,7 @@ struct ArchiveBuilder {
     for (auto &e : this->children) {
       imported.emplace_back(ImportedModKind::GLOBAL, e.build());
     }
-    return std::make_shared<ModuleArchive>(this->id, 0, false, std::move(handles),
+    return std::make_shared<ModuleArchive>(this->id, 0, ModAttr{}, std::move(handles),
                                            std::move(imported));
   }
 };
@@ -717,7 +717,7 @@ struct Builder {
     std::vector<std::pair<ImportedModKind, ModuleArchivePtr>> deps;
     build(deps, std::forward<Args>(args)...);
     auto src = this->srcMan.update(path, 0, "");
-    auto archive = std::make_shared<ModuleArchive>(src->getSrcId(), src->getVersion(), false,
+    auto archive = std::make_shared<ModuleArchive>(src->getSrcId(), src->getVersion(), ModAttr{},
                                                    std::move(handles), std::move(deps));
     this->archives.add(archive);
     return archive;